djangoForClientApp/requirements.txt
Django
djangorestframework
django-cors-headers
djoser
pytest==5.4.1
pytest-django==3.8.0